feat: split streams and filters, create index.js file
