+++ packages/streams/utils/streams.js
@@ -1,82 +1,53 @@
 import * as fs from 'node:fs';
+import { extractWords, getVectorFromWords } from './filters.js';
 
-const space = '\u0020';
 const newLine = '\u000A';
-const nonTextCharacters = ['.', ',', ':', ';', '%'];
 
-const readStream = fs.createReadStream('../files/input-names', {
-  encoding: 'utf8',
-});
-const writeStream = fs.createWriteStream('../files/output-names', {
-  encoding: 'utf8',
-});
+function executeStreams({ inputFilePath, outputFilePath }) {
+  const readStream = fs.createReadStream(inputFilePath, {
+    encoding: 'utf8',
+  });
+  const writeStream = fs.createWriteStream(outputFilePath, {
+    encoding: 'utf8',
+  });
 
-const words = [];
+  const words = [];
 
-readStream.on('data', (chunk) => {
-  words.push(...extractWords(chunk));
-});
+  readStream.on('data', (chunk) => {
+    words.push(...extractWords(chunk));
+  });
 
-readStream.on('end', () => {
-  const sortedWords = [...words].sort();
-  const vector = getVectorFromWords(sortedWords);
+  readStream.on('end', () => {
+    const sortedWords = [...words].sort();
+    const vector = getVectorFromWords(sortedWords);
 
-  const resultWords = words.reduce((res, word) => {
-    return res.concat(`${word}${newLine}`);
-  }, '');
+    const resultWords = words.reduce((res, word) => {
+      return res.concat(`${word}${newLine}`);
+    }, '');
 
-  const resultSortedWords = sortedWords.reduce((res, word) => {
-    return res.concat(`${word}${newLine}`);
-  }, '');
+    const resultSortedWords = sortedWords.reduce((res, word) => {
+      return res.concat(`${word}${newLine}`);
+    }, '');
 
-  const resultVector = Object.values(vector).reduce((res, value) => {
-    return res.concat(`${JSON.stringify(value)}${newLine}`);
-  }, '');
+    const resultVector = Object.values(vector).reduce((res, value) => {
+      return res.concat(`${JSON.stringify(value)}${newLine}`);
+    }, '');
 
-  writeStream.write(`Words array:${newLine}${newLine}`);
-  writeStream.write(resultWords);
-  writeStream.write(newLine);
-  writeStream.write(`Sorted words:${newLine}${newLine}`);
-  writeStream.write(resultSortedWords);
-  writeStream.write(newLine);
-  writeStream.write(`Result vector:${newLine}${newLine}`);
-  writeStream.write(resultVector);
-  writeStream.write(newLine);
-  writeStream.end('Close write stream\n');
-});
+    writeStream.write(`Words array:${newLine}${newLine}`);
+    writeStream.write(resultWords);
+    writeStream.write(newLine);
+    writeStream.write(`Sorted words:${newLine}${newLine}`);
+    writeStream.write(resultSortedWords);
+    writeStream.write(newLine);
+    writeStream.write(`Result vector:${newLine}${newLine}`);
+    writeStream.write(resultVector);
+    writeStream.write(newLine);
+    writeStream.end('Close write stream\n');
+  });
 
-readStream.on('error', (err) => {
-  console.log(err.stack);
-});
-
-function extractWords(chunk) {
-  return [].concat(
-    ...chunk.split(newLine).map((line) => {
-      const filteredLine = filterLine(line);
-      const lineWords = getWordsFromLine(filteredLine);
-      return lineWords;
-    })
-  );
+  readStream.on('error', (err) => {
+    console.log(err.stack);
+  });
 }
 
-function filterLine(line) {
-  return nonTextCharacters.reduce((filteredLine, char) => {
-    return filteredLine.split(char).join('');
-  }, line);
-}
-
-function getWordsFromLine(line) {
-  return line.split(space).filter((word) => Boolean(word));
-}
-
-function getVectorFromWords(words) {
-  return words.reduce((vector, word) => {
-    if (Object.prototype.hasOwnProperty.call(vector, word)) {
-      vector[word].amount += 1;
-    } else {
-      vector[word] = { word, amount: 1 };
-    }
-
-    return vector;
-  }, {});
-}
+export { executeStreams };
